renaming the files and adding more unit tests
